+++ src/t_types/object_sphere.c
@@ -6,12 +6,19 @@
 /*   By: dda-cunh <dda-cunh@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/15 16:04:18 by dda-cunh          #+#    #+#             */
-/*   Updated: 2024/03/30 15:29:39 by dda-cunh         ###   ########.fr       */
+/*   Updated: 2024/05/11 13:15:45 by dda-cunh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/miniRT.h"
 
+/**
+ * Calculates the collision point between a ray and a sphere object.
+ *
+ * @param self	The sphere object.
+ * @param ray	The ray to check for collision.
+ * @return		The collision point structure.
+ */
 static t_coll_point3	collide(t_object_sphere *self, t_ray3 ray)
 {
 	t_point3	coll_coords;
@@ -24,7 +31,7 @@ static t_coll_point3	collide(t_object_sphere *self, t_ray3 ray)
 	b = 2 * vec3_dot_product(ray.direction, vec3_sub(ray.origin,
 				self->center));
 	c = vec3_dot_product(vec3_sub(ray.origin, self->center),
-			vec3_sub(ray.origin, self->center)) - powf(self->diameter / 2, 2);
+			vec3_sub(ray.origin, self->center)) - pow(self->diameter / 2, 2);
 	scalar = quadratic_smallest_pos(a, b, c);
 	if (!valid_collision(scalar))
 		return (get_no_collision());
@@ -40,12 +47,25 @@ static t_coll_point3	collide(t_object_sphere *self, t_ray3 ray)
 		});
 }
 
+/**
+ * Destroys a sphere object.
+ *
+ * @param self	The sphere object to destroy.
+ */
 static void	destroy(t_object_sphere *self)
 {
 	if (self)
 		free(self);
 }
 
+/**
+ * Creates a new sphere object.
+ *
+ * @param center	The center point of the sphere.
+ * @param color		The color of the sphere.
+ * @param diameter	The diameter of the sphere.
+ * @return			The newly created sphere object.
+ */
 t_object_sphere	*new_sphere(t_point3 center, t_color color, double diameter)
 {
 	t_object_sphere	*obj;
@@ -53,7 +73,14 @@ t_object_sphere	*new_sphere(t_point3 center, t_color color, double diameter)
 	obj = ft_calloc(1, sizeof(t_object_sphere));
 	if (!obj)
 		return (NULL);
-	*obj = (t_object_sphere){ID_SPHERE, center, color, diameter, destroy,
-		collide};
+	*obj = (t_object_sphere)
+	{
+		ID_SPHERE,
+		center,
+		color,
+		diameter,
+		destroy,
+		collide
+	};
 	return (obj);
 }
